Update and rename team0.py to team10.py
Fixed the algorithim
